chore: new folder metadata
