+++ src/app/new/clone/configure/[id]/page.tsx
@@ -1,4 +1,5 @@
 import React from 'react'
+import { Metadata } from 'next'
 import { Params } from 'next/dist/shared/lib/router/utils/route-matcher'
 
 import Checkout from '@root/app/new/(checkout)/Checkout'
@@ -10,3 +11,7 @@ const ConfigureDraftFromTemplate: React.FC<{
 }
 
 export default ConfigureDraftFromTemplate
+
+export const metadata: Metadata = {
+  title: 'Configuration | Payload Cloud',
+}
