+++ src/app/new/clone/page.tsx
@@ -1,8 +1,13 @@
 import React from 'react'
 import { NewProjectBlock } from '@blocks/NewProject'
+import { Metadata } from 'next'
 
 const ProjectFromTemplate: React.FC = () => {
   return <NewProjectBlock />
 }
 
 export default ProjectFromTemplate
+
+export const metadata: Metadata = {
+  title: 'Clone Template | Payload Cloud',
+}
