+++ src/app/new/import/page.tsx
@@ -1,43 +1,11 @@
-'use client'
+import { Metadata } from 'next'
 
-import React, { Fragment } from 'react'
+import { ProjectFromImportPage } from './client_page'
 
-import { Breadcrumbs } from '@components/Breadcrumbs'
-import { Gutter } from '@components/Gutter'
-import { Heading } from '@components/Heading'
-import { LoadingShimmer } from '@components/LoadingShimmer'
-import { useGitAuthRedirect } from '../authorize/useGitAuthRedirect'
-import { ImportProject } from './ImportProject'
-
-const title = `Import a codebase`
-
-const ProjectFromImport: React.FC = () => {
-  const { tokenLoading, tokenIsValid } = useGitAuthRedirect()
-
-  return (
-    <Fragment>
-      <Gutter>
-        <div>
-          <Breadcrumbs
-            items={[
-              {
-                label: 'New',
-                url: '/new',
-              },
-              {
-                label: 'Import',
-              },
-            ]}
-          />
-          <Heading marginTop={false} element="h1">
-            {title}
-          </Heading>
-        </div>
-        {tokenLoading && <LoadingShimmer number={3} />}
-      </Gutter>
-      {!tokenLoading && tokenIsValid && <ImportProject />}
-    </Fragment>
-  )
+export default () => {
+  return <ProjectFromImportPage />
 }
 
-export default ProjectFromImport
+export const metadata: Metadata = {
+  title: 'Import Project | Payload Cloud',
+}
